+++ src/app/cart/cart.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { CartListComponent } from './components/cart-list/cart-list.component';
+import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { SharedModule } from '../shared/shared.module';
+import { RouterModule } from '@angular/router';
+
+
+
+@NgModule({
+  declarations: [
+    CartListComponent,
+    CartItemComponent
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+    RouterModule
+  ],
+  exports: [
+    CartListComponent
+  ]
+})
+export class CartModule { }