+++ src/app/products/components/product/product.component.ts
@@ -1,13 +1,26 @@
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+
+import { ProductModel } from '../../models/product-model';
+
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
-  styleUrls: ['./product.component.css']
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProductComponent {
 
+  @Input() product!: ProductModel;
+  @Output() addToCart: EventEmitter<ProductModel> = new EventEmitter();
+
   onAddToCart(): void {
-    console.log('You bought a product!')
+    this.product.isInCart = true;
+    this.addToCart.emit(this.product);
+  }
+
+  getTooltip(): string {
+    return !this.product.isAvailable ? 
+    'Product is unavailable' : this.product.isInCart ? 
+    'Product is already in cart' : 'Add to cart';
   }
 }